requirements.txt
flask==3.0.3
flask-cors==4.0.0
gunicorn==21.2.0
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3
requests==2.32.3
transformers==4.35.2
sqlalchemy==2.0.23
